Reverted change, and moved to ReviewerTest.kt and formated it for it.
