+++ AnkiDroid/src/test/java/com/ichi2/anki/ReviewerTest.kt
@@ -27,6 +27,7 @@ import com.ichi2.anki.AnkiDroidJsAPITest.Companion.getDataFromRequest
 import com.ichi2.anki.AnkiDroidJsAPITest.Companion.jsApiContract
 import com.ichi2.anki.cardviewer.ViewerCommand.FLIP_OR_ANSWER_EASE1
 import com.ichi2.anki.cardviewer.ViewerCommand.MARK
+import com.ichi2.anki.model.CardStateFilter
 import com.ichi2.anki.preferences.PreferenceTestUtils
 import com.ichi2.anki.preferences.sharedPrefs
 import com.ichi2.anki.reviewer.ActionButtonStatus
@@ -42,12 +43,14 @@ import com.ichi2.testutils.MockTime
 import com.ichi2.testutils.OS
 import com.ichi2.utils.KotlinCleanup
 import com.ichi2.utils.deepClone
+import junit.framework.TestCase.assertFalse
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.*
 import org.json.JSONArray
 import org.junit.Ignore
 import org.junit.Test
 import org.junit.runner.RunWith
+import timber.log.Timber
 import kotlin.test.junit5.JUnit5Asserter.assertNotNull
 
 @RunWith(AndroidJUnit4::class)
@@ -64,6 +67,44 @@ class ReviewerTest : RobolectricTest() {
                 )
             }
         }
+    }
+
+    @Test
+    fun testOnSelectedTags() {
+        // Add a note using basic model
+        addNoteUsingBasicModel()
+
+        // Start the Reviewer activity
+        val viewer = startRegularActivity<Reviewer>()
+
+        // Create a list of tags
+        val tags = listOf("tag1", "tag2")
+
+        // Define an arbitrary filter
+        val ARBITRARY_FILTER = CardStateFilter.DUE
+
+        // Assert that currentCard is not null before calling onSelectedTags
+        assertNotNull("currentCard should not be null", viewer.currentCard)
+
+        Timber.d("Test", "Before first call to onSelectedTags")
+
+        // Call onSelectedTags method
+        viewer.onSelectedTags(tags, emptyList(), ARBITRARY_FILTER)
+
+        Timber.d("Test", "After first call to onSelectedTags")
+
+        // Assert that the card is not flipped
+        assertFalse(viewer.isDisplayingAnswer)
+
+        Timber.d("Test", "Before second call to onSelectedTags")
+
+        // Call onSelectedTags method again
+        viewer.onSelectedTags(tags, emptyList(), ARBITRARY_FILTER)
+
+        Timber.d("Test", "After second call to onSelectedTags")
+
+        // Assert that the card is not flipped
+        assertFalse(viewer.isDisplayingAnswer)
     }
 
     @Test
